Split events.test -> events-id.test
